Integrate launchPythonMagnetController into one function
I agree that the way `launchPythonMagnetController` works is not nice. Here's a little workaround. It builds the function name up from the machine mode and area. Of course, it would be better if the magnet controller exposed a catch-all function like `Magnet_Controller(mode, area)` where you could specify `mode` and `area` as strings.