Increased functionality of BPM apps.
There is now a master BPM app, which can launch the other BPM applications. The user can specify which beamline they are interested in, and whether they would like to instantiate a virtual, physical, or offline BPM controller. This folder structure is currently required in order for the master app to communicate with the other apps, but adding in any future applications should be easy.