[parasol] Add linac1 to repertoire
Parasol now knows about Linac-1. This necessitated a slight change to
the way it handles RF fields; now we can consider an array of cells each
of which has a relative phase advance from the last one. (Previously RF
fields were handled as a fieldmap which oscillated all with the same
phase.)

The FORTRAN calcMomentum routine has been updated and recompiled. It was
shamefully uncommented so I've remedied that.